[refactor/TK-148]: dto/response/settlement 폴더 생성
