+++ src/main/java/withbeetravel/service/settlement/SettlementService.java
@@ -1,7 +1,7 @@
 package withbeetravel.service.settlement;
 
 import withbeetravel.dto.response.SuccessResponse;
-import withbeetravel.dto.settlement.ShowSettlementDetailResponse;
+import withbeetravel.dto.response.settlement.ShowSettlementDetailResponse;
 
 public interface SettlementService {
     SuccessResponse<ShowSettlementDetailResponse> getSettlementDetails(Long userId, Long travelId);
