+++ src/main/java/withbeetravel/service/settlement/SettlementServiceImpl.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import withbeetravel.domain.*;
 import withbeetravel.dto.response.SuccessResponse;
-import withbeetravel.dto.settlement.ShowMyDetailPaymentResponse;
-import withbeetravel.dto.settlement.ShowMyTotalPaymentResponse;
-import withbeetravel.dto.settlement.ShowOtherSettlementResponse;
-import withbeetravel.dto.settlement.ShowSettlementDetailResponse;
+import withbeetravel.dto.response.settlement.ShowMyDetailPaymentResponse;
+import withbeetravel.dto.response.settlement.ShowMyTotalPaymentResponse;
+import withbeetravel.dto.response.settlement.ShowOtherSettlementResponse;
+import withbeetravel.dto.response.settlement.ShowSettlementDetailResponse;
 import withbeetravel.exception.CustomException;
 import withbeetravel.exception.error.AuthErrorCode;
 import withbeetravel.exception.error.SettlementErrorCode;
